mcs_km: modify the pagetable phyaddr and add boot table
description: The old boot and page table share a partition, but this
partition will be used by linux, resulting in the page table be cleared.
Now the boot and page table partition is separated. The boot partition
use the reserved low address 0x0, and the page table partition use the
reserved address 0xa0000.

Signed-off-by: lishun <lishun27@huawei.com>