fix: shared memory pool size for vring buffer
In the MICA implementation based on openAMP, we have dedicated data
structures for storing vring info.The dedicated data structures are
in the end of the shared memory region.

However, when initializing the shared memory pool for allocating
RPMsg buffer, we forgot to exclude the dedicated data structures
from the total pool length. This submission fixes this problem.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>