feat: debug RTOS based on MICA (the commonly used functions and definitions)
- define max buffer length according to the message length from RTOS.

- define default message priority for message queue.

- the log function used to write logs into logfile.txt whose path is defined through
LOG_FILE_PATH.

- the conversion macros for transferring between (void *) type and int
type.

- fix the compilation error for not defining "uint32_t" in mcs_common.h.

Since we need to transfer messages, many modules need buffers.
The max message length generated by RTOS is based on the number
and length of registers, which is no more than 1600, so we do
not need to allocate more spaces.

For now all the messages transferred have the same priority,
so we just choose the lowest priority number 0 as the default priority.

The logs will be stored into pre-defined path and the format is:
[time stamp] [module name] error msg

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>