feat: debug RTOS based on MICA (ring buffer)
- the ring_buffer.h file contains the definitions and operations
of the ring buffer.

- the mica_debug_ring_buffer part of data contains the data structure
and interfaces for sending and receiving messages to and from the proxy
server module. This part of code calls the functions in ring_buffer.h
to directly manipulate the shared memory.

The reason why we have independent files for ring buffer
operations is because it is easy for us to extend the software
architecture to support other communication mechanisms based on
shared memory.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>