feat!: debug RTOS based on MICA (main file)
- it provides a unified entrance for using pty and debugging
  applications. While using the debugging application, it will
  automatically enter the GDB command line and let users to
  interact with the RTOS.
  If users quit from the GDB command line, the whole debugging
  application along with the remote RTOS will exit as well.

- the CMakeList.txt contains the configuration of compiling and
  linking to build up the debuging application.

BREAKING CHANGE: if users want to compile a userspace program which
is capable of debugging the RTOS through MICA framework, they should
set the macro "DEMO_TARGET" to be "mica_demo". Meanwhile,
they must also point out the communication technology based on
shared memory they want to use. Define macro "CONFIG_RING_BUFFER"
stands for using ring buffer, and define macro
"CONFIG_VIRTIO_RPMSG" stands for using RPMsg based on
virtio device (this is an incoming feature). The two macros are
mutually exclusive. In other words, they cannot be defined at the
same time, otherwise in default we use the ring buffer as the
shared memory communication mechanism.

If the users want to use the previous pty application, they can just
start the client os, open the pty device as previous.
If the users want to debug the RTOS,
they need to boot up the RTOS with an extra flag "-d" with the
path to the ELF executable containing debugging symbols.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>