test_ring_buffer: only compile when specifying TEST_RING_BUFFER
For ordinary use, we do not need the testcase for ring buffer and
debugging module. When debugging micad, now the micad
allocates memory dynamically for the ring buffer, and store this
infomation in resource table. However, the testcase
"test_ring_buffer" cannot access the resource table.
Thus, we need to statically allocate an area of memory for
the ring buffer, so that both micad and the testcase can
communicate based on a predefined static memory area.

As a result, if we want to debug the functionality of
current debugging module, we should define "TEST_RING_BUFFER"
to compile the testcase and replace "setup_rbuf_dev" with
"setup_test_rbuf" to use a predefined area of memory for ring
buffer. For ordinary use, just do not define this macro, and
the testcase will not be generated, while the ring buffer device
of micad will allocate memory for ring buffer dynamically.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>