mica gdb: do not resend packet when timeout
When sending a command and do not receive a feedback from the
remote for a long time, the gdb client will resend the command
again. However, after stopping and starting the uniproton with
gdbstub, it cannot be soonly stopped again, because it is busy
waiting the gdb commands. Considering that it has very low
possibility to lost packet between the GDB client and MICA
GDB server, we set remote timeout to be unlimited, so that
each command will be sent only once.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>