services/debug: fix segment fault when stopping MICA client
On hi3093, when executing "run" command, it usually causes a
segment fault.

The reason is, when restarting MICA client, the thread for sending
messages to RTOS through ring buffer module has not been cancelled,
so it is still using the shared memory for communication. However,
the client has been stopped and shared memory has been freed. Thus,
memory access causes segment fault.

The solution is to wait until the threads been cancelled.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>