services/debug: a workaround solving not receivng interrupts
If the user connects to UniProton through GDB stub and input
"run" command immediately, the UniProton may not initialize
the interrupt controller, so it cannot receive interrupt.
As a workaround, we mandatorily sleep for 1 second to ensure
the UniProton initializing the interrupt controller.

Note: this is only a workaround. Firstly, 1 second is from
experience, it may not need 1 second to initialize the interrupt
controller, or under some circumstances it needs more than 1
second to initialize. It is better for us to have a
synchronization mechanism to know if the remote has received the
interrupt or not.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>