mica: change the path in debugging info to be relative
Previously, we compile the "micad" on compilation machine, and
copy it to the host machine for debugging. On host machine, the
path to the MCS source code is different from the path on
compilation machine. The absolute path is hard for us to debug,
because the GDB will find the source code according to the path
containing the absolute path.

Thus, replace the path to source code with relative path, and use
the "directory" command in GDB to define the path to MCS repo on
host machine, so that the "list" command can list the source code
correctly.

Signed-off-by: Yongmao Luo <luoyongmao@huawei.com>